+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Fixed
+
+- Quote table prefix and name correctly in `Carbonite.Migrations` functions. (@ravensiris)
+
 ## [0.14.2] - 2024-08-21
 
 **New migration patches:** 11